+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## 🌟 Features
 
 - You can now paste images directly to the editor
+- You can now drop image files directly. You can drop more than one image and all of them will be inserted to the page.
 - Editor now has a context menu
 - Added word count - you can view it by clicking the menu icon on the top right.
 - Added translations for Turkish